design: 달력 아이콘을 종이비행기 아이콘으로 교체
