--- src/components/icons/calendar.tsx
@@ -1,35 +1,0 @@
-interface IconProps {
-  width?: number;
-  height?: number;
-  color?: string;
-}
-
-export const CalendarIcon = ({ width = 16, height = 16 }: IconProps) => {
-  return (
-    <svg
-      width={width}
-      height={height}
-      viewBox="0 0 16 16"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        opacity="0.5"
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M4.8335 4.17184V4.33317C4.8335 4.46578 4.88617 4.59296 4.97994 4.68672C5.07371 4.78049 5.20089 4.83317 5.3335 4.83317C5.4661 4.83317 5.59328 4.78049 5.68705 4.68672C5.78082 4.59296 5.8335 4.46578 5.8335 4.33317V4.1665H10.1668V4.33317C10.1668 4.46578 10.2195 4.59296 10.3133 4.68672C10.407 4.78049 10.5342 4.83317 10.6668 4.83317C10.7994 4.83317 10.9266 4.78049 11.0204 4.68672C11.1142 4.59296 11.1668 4.46578 11.1668 4.33317V4.17184C11.2406 4.1745 11.3077 4.17828 11.3682 4.18317C11.5602 4.1985 11.6515 4.2265 11.7115 4.25717C11.8688 4.33717 11.9962 4.4645 12.0762 4.62184C12.1068 4.68184 12.1348 4.77317 12.1502 4.9645C12.1662 5.16184 12.1668 5.41784 12.1668 5.79984V6.1665H3.8335V5.79984C3.8335 5.4185 3.8335 5.16184 3.85016 4.9645C3.8655 4.77317 3.8935 4.68184 3.92416 4.62184C4.0041 4.46478 4.13177 4.33711 4.28883 4.25717C4.34883 4.2265 4.44016 4.1985 4.6315 4.18317C4.69872 4.17769 4.76608 4.17391 4.8335 4.17184ZM7.00016 4.6665C6.86755 4.6665 6.74038 4.71918 6.64661 4.81295C6.55284 4.90672 6.50016 5.0339 6.50016 5.1665C6.50016 5.29911 6.55284 5.42629 6.64661 5.52006C6.74038 5.61383 6.86755 5.6665 7.00016 5.6665H9.00016C9.13277 5.6665 9.25995 5.61383 9.35372 5.52006C9.44748 5.42629 9.50016 5.29911 9.50016 5.1665C9.50016 5.0339 9.44748 4.90672 9.35372 4.81295C9.25995 4.71918 9.13277 4.6665 9.00016 4.6665H7.00016Z"
-        fill="black"
-      />
-      <path
-        d="M5.66683 9.33317C5.88784 9.33317 6.0998 9.24537 6.25609 9.08909C6.41237 8.93281 6.50016 8.72085 6.50016 8.49984C6.50016 8.27882 6.41237 8.06686 6.25609 7.91058C6.0998 7.7543 5.88784 7.6665 5.66683 7.6665C5.44582 7.6665 5.23385 7.7543 5.07757 7.91058C4.92129 8.06686 4.8335 8.27882 4.8335 8.49984C4.8335 8.72085 4.92129 8.93281 5.07757 9.08909C5.23385 9.24537 5.44582 9.33317 5.66683 9.33317ZM5.66683 11.6665C5.88784 11.6665 6.0998 11.5787 6.25609 11.4224C6.41237 11.2661 6.50016 11.0542 6.50016 10.8332C6.50016 10.6122 6.41237 10.4002 6.25609 10.2439C6.0998 10.0876 5.88784 9.99984 5.66683 9.99984C5.44582 9.99984 5.23385 10.0876 5.07757 10.2439C4.92129 10.4002 4.8335 10.6122 4.8335 10.8332C4.8335 11.0542 4.92129 11.2661 5.07757 11.4224C5.23385 11.5787 5.44582 11.6665 5.66683 11.6665ZM8.8335 8.49984C8.8335 8.72085 8.7457 8.93281 8.58942 9.08909C8.43314 9.24537 8.22118 9.33317 8.00016 9.33317C7.77915 9.33317 7.56719 9.24537 7.41091 9.08909C7.25463 8.93281 7.16683 8.72085 7.16683 8.49984C7.16683 8.27882 7.25463 8.06686 7.41091 7.91058C7.56719 7.7543 7.77915 7.6665 8.00016 7.6665C8.22118 7.6665 8.43314 7.7543 8.58942 7.91058C8.7457 8.06686 8.8335 8.27882 8.8335 8.49984ZM8.00016 11.6665C8.22118 11.6665 8.43314 11.5787 8.58942 11.4224C8.7457 11.2661 8.8335 11.0542 8.8335 10.8332C8.8335 10.6122 8.7457 10.4002 8.58942 10.2439C8.43314 10.0876 8.22118 9.99984 8.00016 9.99984C7.77915 9.99984 7.56719 10.0876 7.41091 10.2439C7.25463 10.4002 7.16683 10.6122 7.16683 10.8332C7.16683 11.0542 7.25463 11.2661 7.41091 11.4224C7.56719 11.5787 7.77915 11.6665 8.00016 11.6665ZM11.1668 8.49984C11.1668 8.72085 11.079 8.93281 10.9228 9.08909C10.7665 9.24537 10.5545 9.33317 10.3335 9.33317C10.1125 9.33317 9.90052 9.24537 9.74424 9.08909C9.58796 8.93281 9.50016 8.72085 9.50016 8.49984C9.50016 8.27882 9.58796 8.06686 9.74424 7.91058C9.90052 7.7543 10.1125 7.6665 10.3335 7.6665C10.5545 7.6665 10.7665 7.7543 10.9228 7.91058C11.079 8.06686 11.1668 8.27882 11.1668 8.49984Z"
-        fill="black"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M5.3335 2.1665C5.4661 2.1665 5.59328 2.21918 5.68705 2.31295C5.78082 2.40672 5.8335 2.5339 5.8335 2.6665V3.1665H10.1668V2.6665C10.1668 2.5339 10.2195 2.40672 10.3133 2.31295C10.407 2.21918 10.5342 2.1665 10.6668 2.1665C10.7994 2.1665 10.9266 2.21918 11.0204 2.31295C11.1142 2.40672 11.1668 2.5339 11.1668 2.6665V3.17184C11.2682 3.1745 11.3626 3.17939 11.4502 3.1865C11.7035 3.2065 11.9408 3.25184 12.1655 3.3665C12.5105 3.54228 12.791 3.8228 12.9668 4.16784C13.0815 4.3925 13.1268 4.62984 13.1468 4.88317C13.1668 5.1265 13.1668 5.42317 13.1668 5.77984V10.8865C13.1668 11.2432 13.1668 11.5398 13.1468 11.7832C13.1268 12.0365 13.0815 12.2738 12.9668 12.4985C12.7912 12.8434 12.511 13.1239 12.1662 13.2998C11.9408 13.4145 11.7035 13.4598 11.4502 13.4798C11.2068 13.4998 10.9102 13.4998 10.5542 13.4998H5.44683C5.09016 13.4998 4.7935 13.4998 4.55016 13.4798C4.29683 13.4598 4.0595 13.4145 3.83483 13.2998C3.49 13.1244 3.2095 12.8444 3.0335 12.4998C2.91883 12.2745 2.8735 12.0372 2.8535 11.7838C2.8335 11.5405 2.8335 11.2438 2.8335 10.8878V5.77984C2.8335 5.42317 2.8335 5.1265 2.8535 4.88317C2.8735 4.62984 2.91883 4.3925 3.0335 4.16784C3.20928 3.8228 3.48979 3.54228 3.83483 3.3665C4.0595 3.25184 4.29683 3.2065 4.55016 3.1865C4.63772 3.17939 4.73216 3.1745 4.8335 3.17184V2.6665C4.8335 2.60084 4.84643 2.53582 4.87156 2.47516C4.89668 2.4145 4.93351 2.35938 4.97994 2.31295C5.02637 2.26652 5.08149 2.22969 5.14215 2.20456C5.20282 2.17944 5.26784 2.1665 5.3335 2.1665ZM4.8335 4.33317V4.17184C4.76608 4.17391 4.69872 4.17769 4.6315 4.18317C4.44016 4.1985 4.34883 4.2265 4.28883 4.25717C4.13177 4.33711 4.0041 4.46478 3.92416 4.62184C3.8935 4.68184 3.8655 4.77317 3.85016 4.9645C3.83416 5.16184 3.8335 5.41784 3.8335 5.79984V6.1665H12.1668V5.79984C12.1668 5.4185 12.1668 5.16184 12.1502 4.9645C12.1348 4.77317 12.1068 4.68184 12.0762 4.62184C11.9962 4.46478 11.8686 4.33711 11.7115 4.25717C11.6515 4.2265 11.5602 4.1985 11.3682 4.18317C11.3012 4.1777 11.234 4.17392 11.1668 4.17184V4.33317C11.1668 4.46578 11.1142 4.59296 11.0204 4.68672C10.9266 4.78049 10.7994 4.83317 10.6668 4.83317C10.5342 4.83317 10.407 4.78049 10.3133 4.68672C10.2195 4.59296 10.1668 4.46578 10.1668 4.33317V4.1665H5.8335V4.33317C5.8335 4.46578 5.78082 4.59296 5.68705 4.68672C5.59328 4.78049 5.4661 4.83317 5.3335 4.83317C5.20089 4.83317 5.07371 4.78049 4.97994 4.68672C4.88617 4.59296 4.8335 4.46578 4.8335 4.33317ZM12.1668 6.83317H3.8335V10.8665C3.8335 11.2478 3.8335 11.5045 3.85016 11.7012C3.8655 11.8932 3.8935 11.9845 3.92416 12.0445C4.00416 12.2018 4.1315 12.3292 4.28883 12.4092C4.34883 12.4398 4.44016 12.4678 4.6315 12.4832C4.82883 12.4992 5.08483 12.4998 5.46683 12.4998H10.5335C10.9148 12.4998 11.1715 12.4998 11.3682 12.4832C11.5602 12.4678 11.6515 12.4398 11.7115 12.4092C11.8686 12.3292 11.9962 12.2016 12.0762 12.0445C12.1068 11.9845 12.1348 11.8932 12.1502 11.7012C12.1662 11.5045 12.1668 11.2478 12.1668 10.8665V6.83317Z"
-        fill="black"
-      />
-    </svg>
-  );
-};
