+++ src/components/icons/index.ts
@@ -3,5 +3,5 @@ export * from "./check-box";
 export * from "./light-bulb";
 export * from "./warning";
 export * from "./star";
-export * from "./calendar";
 export * from "./right-arrow";
+export * from "./paper-plane";
