+++ src/components/icons/paper-plane.tsx
@@ -1,0 +1,51 @@
+import { css } from "#/styled-system/css";
+
+interface IconProps {
+  width?: number;
+  height?: number;
+  color?: string;
+}
+
+export const PaperPlaneIcon = ({ width = 14, height = 14 }: IconProps) => {
+  return (
+    <svg
+      width={width}
+      height={height}
+      viewBox="0 0 14 14"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <g clipPath="url(#clip0_579_195)">
+        <path
+          className={iconStyle}
+          d="M1.06656 12.9324C1.15602 13.0218 1.26944 13.0835 1.39314 13.11C1.51685 13.1365 1.64557 13.1266 1.76382 13.0816L12.7021 8.94153H12.7043C12.8305 8.89302 12.9385 8.80677 13.0138 8.69454C13.089 8.58231 13.1278 8.44959 13.1248 8.3145C13.1218 8.1794 13.0772 8.04853 12.9971 7.93974C12.9169 7.83096 12.8051 7.74957 12.677 7.70669L12.6658 7.70313L8.91148 6.09997C8.83825 6.07774 8.76048 6.07511 8.68591 6.09233C8.61135 6.10956 8.54261 6.14604 8.48656 6.19813L2.46109 11.8126C2.44313 11.8306 2.42182 11.8448 2.39836 11.8545C2.3749 11.8643 2.34976 11.8693 2.32437 11.8693C2.29898 11.8693 2.27384 11.8643 2.25038 11.8545C2.22692 11.8448 2.2056 11.8306 2.18765 11.8126C2.1697 11.7947 2.15545 11.7734 2.14574 11.7499C2.13602 11.7264 2.13102 11.7013 2.13102 11.6759C2.13102 11.6505 2.13602 11.6254 2.14574 11.6019C2.15545 11.5785 2.1697 11.5571 2.18765 11.5392L7.80242 5.51645C7.85451 5.4604 7.89099 5.39166 7.90822 5.3171C7.92544 5.24253 7.92281 5.16476 7.90058 5.09153L6.29687 1.33505C6.29523 1.32958 6.29359 1.32466 6.29167 1.31946C6.20417 1.06599 5.98269 0.887157 5.71499 0.875126H5.68765C5.55251 0.874355 5.42026 0.914258 5.3081 0.98965C5.19594 1.06504 5.10906 1.17243 5.05874 1.29786L0.919173 12.2332C0.873561 12.3515 0.863207 12.4805 0.889363 12.6046C0.915519 12.7287 0.977064 12.8425 1.06656 12.9324Z"
+        />
+      </g>
+      <defs>
+        <clipPath id="clip0_579_195">
+          <rect
+            className={iconStyle}
+            width="14"
+            height="14"
+            transform="matrix(-1 0 0 -1 14 14)"
+          />
+        </clipPath>
+      </defs>
+    </svg>
+
+    // <svg
+    //   width={width}
+    //   height={height}
+    //   viewBox="0 0 16 16"
+    //   fill="none"
+    //   xmlns="http://www.w3.org/2000/svg"
+    // >
+    //   <path
+    //     className={iconStyle}
+    //     d="M1.21892 14.7799C1.32117 14.8821 1.45078 14.9526 1.59216 14.9829C1.73354 15.0132 1.88066 15.0019 2.0158 14.9505L14.5167 10.219H14.5192C14.6634 10.1635 14.7869 10.0649 14.8729 9.93669C14.9589 9.80843 15.0032 9.65674 14.9998 9.50235C14.9964 9.34796 14.9454 9.19838 14.8538 9.07406C14.7622 8.94973 14.6344 8.85672 14.488 8.80771L14.4752 8.80365L10.1845 6.97146C10.1009 6.94606 10.012 6.94305 9.92676 6.96273C9.84154 6.98242 9.76299 7.02411 9.69892 7.08365L2.81267 13.5002C2.79215 13.5207 2.76779 13.537 2.74098 13.5481C2.71417 13.5592 2.68544 13.5649 2.65642 13.5649C2.6274 13.5649 2.59867 13.5592 2.57186 13.5481C2.54505 13.537 2.52069 13.5207 2.50017 13.5002C2.47965 13.4797 2.46338 13.4553 2.45227 13.4285C2.44117 13.4017 2.43545 13.373 2.43545 13.344C2.43545 13.3149 2.44117 13.2862 2.45227 13.2594C2.46338 13.2326 2.47965 13.2082 2.50017 13.1877L8.91705 6.30459C8.97658 6.24052 9.01828 6.16197 9.03796 6.07675C9.05765 5.99154 9.05464 5.90265 9.02923 5.81896L7.19642 1.52584C7.19455 1.51959 7.19267 1.51396 7.19048 1.50803C7.09048 1.21834 6.83736 1.01396 6.53142 1.00021H6.50017C6.34572 0.999332 6.19458 1.04494 6.0664 1.1311C5.93822 1.21726 5.83892 1.33999 5.78142 1.48334L1.05048 13.9808C0.998355 14.1161 0.986522 14.2635 1.01641 14.4053C1.04631 14.5471 1.11664 14.6772 1.21892 14.7799Z"
+    //   />
+    // </svg>
+  );
+};
+
+const iconStyle = css({ fill: "background.reverse" });
