+++ src/components/post-list/list-item.tsx
@@ -2,7 +2,7 @@ import { css, cx } from "#/styled-system/css";
 import { pretendard } from "@/styles/font";
 import Image from "next/image";
 import Link from "next/link";
-import { CalendarIcon } from "../icons";
+import { PaperPlaneIcon } from "../icons";
 import { Badge } from "./badge";
 import { convertTagName, TagType } from "@/lib/utils";
 
@@ -48,7 +48,7 @@ export const PostListItem = ({
           </div>
 
           <div className={dateWrapperStyle}>
-            <CalendarIcon />
+            <PaperPlaneIcon />
             <p className={dateStyle}>{createdAt}</p>
           </div>
         </div>
